Fixed LObject::Align

+++ src/LObject.cpp
@@ -23,7 +23,7 @@ void LObject::SetSize(uint16_t x, uint16_t y) {
     lv_obj_set_size(ptr, x, y);
 }
 
-void LObject::Align(lv_align_t alignment, uint16_t x, uint16_t y) {
+void LObject::Align(lv_align_t alignment, int32_t x, int32_t y) {
     lv_obj_align(ptr, alignment, x, y);
 }
