+++ src/LObject.h
@@ -18,7 +18,7 @@ public:
 
     void SetPos(uint16_t x, uint16_t y);
     void SetSize(uint16_t x, uint16_t y);
-    void Align(lv_align_t alignment, uint16_t x, uint16_t y);
+    void Align(lv_align_t alignment, int32_t x, int32_t y);
 
     void AddCallback(Callback callback, lv_event_code_t event);
     void OnPress(Callback callback);
